refactor to separate consumer from aws client
