+++ consumers/aws.go
@@ -2,6 +2,7 @@ package consumers
 
 import (
 	"errors"
+	"fmt"
 
 	"gopkg.in/alecthomas/kingpin.v2"
 
@@ -16,43 +17,45 @@ import (
 	awsclient "github.com/zalando-incubator/mate/pkg/aws"
 )
 
-// Implementations provide access to AWS Route53 API's
-// required calls.
 type AWSClient interface {
 	ListRecordSets(zoneID string) ([]*route53.ResourceRecordSet, error)
 	ChangeRecordSets(upsert, del, create []*route53.ResourceRecordSet, zoneID string) error
-	EndpointsToAlias(endpoints []*pkg.Endpoint) ([]*route53.ResourceRecordSet, error)
-	RecordInfo(records []*route53.ResourceRecordSet) map[string]*pkg.RecordInfo
-	GetGroupID() string
-	GetAssignedTXTRecordObject(record *route53.ResourceRecordSet) *route53.ResourceRecordSet
+	GetCanonicalZoneIDs(lbDNS []string) (map[string]string, error)
 	GetHostedZones() (map[string]string, error)
 }
 
-type awsClient struct {
-	client AWSClient
+type awsConsumer struct {
+	groupID string
+	client  AWSClient
 }
+
+const (
+	evaluateTargetHealth = true
+	defaultTxtTTL        = int64(300)
+)
 
 func init() {
 	kingpin.Flag("aws-record-group-id", "Identifier to filter the mate records ").StringVar(&params.awsGroupID)
 }
 
-// NewAWSRoute53 reates a Consumer instance to sync and process DNS
+// NewAWSConsumer reates a Consumer instance to sync and process DNS
 // entries in AWS Route53.
-func NewAWSRoute53() (Consumer, error) {
+func NewAWSConsumer() (Consumer, error) {
 	if params.awsGroupID == "" {
 		return nil, errors.New("please provide --aws-record-group-id")
 	}
-	return withClient(awsclient.New(awsclient.Options{
-		GroupID: params.awsGroupID,
-	})), nil
+	return withClient(awsclient.New(awsclient.Options{}), params.awsGroupID), nil
 }
 
-func withClient(c AWSClient) *awsClient {
-	return &awsClient{c}
+func withClient(c AWSClient, groupID string) *awsConsumer {
+	return &awsConsumer{
+		groupID: groupID,
+		client:  c,
+	}
 }
 
-func (a *awsClient) Sync(endpoints []*pkg.Endpoint) error {
-	newAliasRecords, err := a.client.EndpointsToAlias(endpoints)
+func (a *awsConsumer) Sync(endpoints []*pkg.Endpoint) error {
+	newAliasRecords, err := a.EndpointsToAlias(endpoints)
 	if err != nil {
 		return err
 	}
@@ -93,13 +96,13 @@ func (a *awsClient) Sync(endpoints []*pkg.Endpoint) error {
 	return nil
 }
 
-func (a *awsClient) syncPerHostedZone(newAliasRecords []*route53.ResourceRecordSet, zoneID string) error {
+func (a *awsConsumer) syncPerHostedZone(newAliasRecords []*route53.ResourceRecordSet, zoneID string) error {
 	existingRecords, err := a.client.ListRecordSets(zoneID)
 	if err != nil {
 		return err
 	}
 
-	recordInfoMap := a.client.RecordInfo(existingRecords)
+	recordInfoMap := a.RecordInfo(existingRecords)
 
 	var upsert, del []*route53.ResourceRecordSet
 
@@ -108,19 +111,19 @@ func (a *awsClient) syncPerHostedZone(newAliasRecords []*route53.ResourceRecordS
 		existingRecordInfo, exist := recordInfoMap[aws.StringValue(newAliasRecord.Name)]
 
 		if !exist { //record does not exist, create it
-			newTXTRecord := a.client.GetAssignedTXTRecordObject(newAliasRecord)
+			newTXTRecord := a.GetAssignedTXTRecordObject(newAliasRecord)
 			upsert = append(upsert, newAliasRecord, newTXTRecord)
 			continue
 		}
 
-		if existingRecordInfo.GroupID != a.client.GetGroupID() { // there exist a record with a different or empty group ID
+		if existingRecordInfo.GroupID != a.GetGroupID() { // there exist a record with a different or empty group ID
 			log.Warnf("Skipping record %s: with a group ID: %s", aws.StringValue(newAliasRecord.Name), existingRecordInfo.GroupID)
 			continue
 		}
 
 		// make sure record only updated when target changes, not to spam AWS route53 API with dummy updates
 		if pkg.SanitizeDNSName(existingRecordInfo.Target) != aws.StringValue(newAliasRecord.AliasTarget.DNSName) {
-			newTXTRecord := a.client.GetAssignedTXTRecordObject(newAliasRecord)
+			newTXTRecord := a.GetAssignedTXTRecordObject(newAliasRecord)
 			upsert = append(upsert, newAliasRecord, newTXTRecord)
 		}
 	}
@@ -128,7 +131,7 @@ func (a *awsClient) syncPerHostedZone(newAliasRecords []*route53.ResourceRecordS
 	//find records to be removed
 	for _, existingRecord := range existingRecords {
 		recordInfo := recordInfoMap[aws.StringValue(existingRecord.Name)]
-		if recordInfo.GroupID == a.client.GetGroupID() {
+		if recordInfo.GroupID == a.GetGroupID() {
 			remove := true
 			for _, newAliasRecord := range newAliasRecords {
 				if aws.StringValue(newAliasRecord.Name) == aws.StringValue(existingRecord.Name) {
@@ -151,7 +154,7 @@ func (a *awsClient) syncPerHostedZone(newAliasRecords []*route53.ResourceRecordS
 	return nil
 }
 
-func (a *awsClient) Consume(endpoints <-chan *pkg.Endpoint, errors chan<- error, done <-chan struct{}, wg *sync.WaitGroup) {
+func (a *awsConsumer) Consume(endpoints <-chan *pkg.Endpoint, errors chan<- error, done <-chan struct{}, wg *sync.WaitGroup) {
 	wg.Add(1)
 	defer wg.Done()
 
@@ -178,18 +181,18 @@ func (a *awsClient) Consume(endpoints <-chan *pkg.Endpoint, errors chan<- error,
 	}
 }
 
-func (a *awsClient) Process(endpoint *pkg.Endpoint) error {
+func (a *awsConsumer) Process(endpoint *pkg.Endpoint) error {
 	hostedZonesMap, err := a.client.GetHostedZones()
 	if err != nil {
 		return err
 	}
 
-	aliasRecords, err := a.client.EndpointsToAlias([]*pkg.Endpoint{endpoint})
+	aliasRecords, err := a.EndpointsToAlias([]*pkg.Endpoint{endpoint})
 	if err != nil {
 		return err
 	}
 
-	create := []*route53.ResourceRecordSet{aliasRecords[0], a.client.GetAssignedTXTRecordObject(aliasRecords[0])}
+	create := []*route53.ResourceRecordSet{aliasRecords[0], a.GetAssignedTXTRecordObject(aliasRecords[0])}
 
 	zoneID := getZoneIDForEndpoint(hostedZonesMap, aliasRecords[0])
 	if zoneID == "" {
@@ -217,4 +220,97 @@ func getZoneIDForEndpoint(hostedZonesMap map[string]string, record *route53.Reso
 		}
 	}
 	return match
+}
+
+//GetGroupID returns the idenitifier for AWS records as stored in TXT records
+func (a *awsConsumer) GetGroupID() string {
+	return fmt.Sprintf("\"mate:%s\"", a.groupID)
+}
+
+//GetAssignedTXTRecordObject returns the TXT record which accompanies the Alias record
+func (a *awsConsumer) GetAssignedTXTRecordObject(aliasRecord *route53.ResourceRecordSet) *route53.ResourceRecordSet {
+	return &route53.ResourceRecordSet{
+		Type: aws.String("TXT"),
+		Name: aliasRecord.Name,
+		TTL:  aws.Int64(defaultTxtTTL),
+		ResourceRecords: []*route53.ResourceRecord{{
+			Value: aws.String(a.GetGroupID()),
+		}},
+	}
+}
+
+//RecordInfo returns the map of record assigned dns to its target and groupID (can be empty)
+func (a *awsConsumer) RecordInfo(records []*route53.ResourceRecordSet) map[string]*pkg.RecordInfo {
+	groupIDMap := map[string]string{} //maps dns to group ID
+
+	for _, record := range records {
+		if (aws.StringValue(record.Type)) == "TXT" {
+			groupIDMap[aws.StringValue(record.Name)] = aws.StringValue(record.ResourceRecords[0].Value)
+		} else {
+			if _, exist := groupIDMap[aws.StringValue(record.Name)]; !exist {
+				groupIDMap[aws.StringValue(record.Name)] = ""
+			}
+		}
+	}
+
+	infoMap := map[string]*pkg.RecordInfo{} //maps record DNS to its GroupID (if exists) and Target (LB)
+	for _, record := range records {
+		groupID := groupIDMap[aws.StringValue(record.Name)]
+		if _, exist := infoMap[aws.StringValue(record.Name)]; !exist {
+			infoMap[aws.StringValue(record.Name)] = &pkg.RecordInfo{
+				GroupID: groupID,
+			}
+		}
+		if aws.StringValue(record.Type) != "TXT" {
+			infoMap[aws.StringValue(record.Name)].Target = getRecordTarget(record)
+		}
+	}
+
+	return infoMap
+}
+
+func getRecordTarget(r *route53.ResourceRecordSet) string {
+	if aws.StringValue(r.Type) == "TXT" {
+		return ""
+	}
+	if r.AliasTarget != nil {
+		return aws.StringValue(r.AliasTarget.DNSName)
+	}
+	return aws.StringValue(r.ResourceRecords[0].Value)
+}
+
+//EndpointsToAlias converts pkg Endpoint to route53 Alias Records
+func (c *awsConsumer) EndpointsToAlias(endpoints []*pkg.Endpoint) ([]*route53.ResourceRecordSet, error) {
+	lbDNS := make([]string, len(endpoints))
+	for i := range endpoints {
+		lbDNS[i] = endpoints[i].Hostname
+	}
+	zoneIDs, err := c.client.GetCanonicalZoneIDs(lbDNS)
+	if err != nil {
+		return nil, err
+	}
+	var rset []*route53.ResourceRecordSet
+
+	for _, ep := range endpoints {
+		if loadBalancerZoneID, exist := zoneIDs[ep.Hostname]; exist {
+			rset = append(rset, c.endpointToAlias(ep, aws.String(loadBalancerZoneID)))
+		} else {
+			log.Errorf("Canonical Zone ID for endpoint: %s is not found", ep.Hostname)
+		}
+	}
+	return rset, nil
+}
+
+//endpointToAlias convert endpoint to an AWS A Alias record
+func (c *awsConsumer) endpointToAlias(ep *pkg.Endpoint, canonicalZoneID *string) *route53.ResourceRecordSet {
+	rs := &route53.ResourceRecordSet{
+		Type: aws.String("A"),
+		Name: aws.String(pkg.SanitizeDNSName(ep.DNSName)),
+		AliasTarget: &route53.AliasTarget{
+			DNSName:              aws.String(pkg.SanitizeDNSName(ep.Hostname)),
+			EvaluateTargetHealth: aws.Bool(evaluateTargetHealth),
+			HostedZoneId:         canonicalZoneID,
+		},
+	}
+	return rs
 }
